Updated to use gcc4.7 llvm_gcc4.2 and clang. Fix warning produced by gcc 4.7

+++ samson/apps/unitTest/lib/au/DescriptorsTest.cpp
@@ -13,6 +13,6 @@ TEST(au_Descriptors, basic) {
   descriptors.Add("Adios");
   descriptors.Add("Adios");
 
-  EXPECT_EQ(2, descriptors.size());
+  EXPECT_EQ((size_t)2, descriptors.size());
   EXPECT_EQ("3 X Adios 2 X Hola", descriptors.str());
 }
